searchfunctioninuser and emailsending in payment and card details added

+++ Backend/User-microservice/src/main/java/com/cg/controller/UserController.java
@@ -27,7 +27,9 @@ public class UserController {
 
 	@Autowired
 	private UserServiceInterface userService;
-
+	/* "firstName":"Tejaswi","lastName":"Midgule","userName":"Tejaswi27","password":"Tejaswi123","mobileNo":"9011872311","email":"tmidgule98@gmail.com","gender":"female",
+	 * "role":"User","dateOfBirth":"1998-11-27"
+	 */
 	//http://localhost:9200/user/addUser
 	@RequestMapping(value = "/addUser", method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
 	public User addUser(@Valid @RequestBody User user) {
@@ -44,19 +46,19 @@ public class UserController {
 
 	//http://localhost:9200/user/deleteUser/12
 	@DeleteMapping(value = "/deleteUser/{userId}")
-	public void deleteUser(@PathVariable Integer userId) {
+	public void deleteUser(@PathVariable Long userId) {
 		this.userService.deleteUser(userId);
 	}
 
 	//http://localhost:9200/user/searchUser/1
 	@GetMapping(value = "/searchUser/{userId}")
-	public User searchUserById(@PathVariable Integer userId) {
+	public User searchUserById(@PathVariable Long userId) {
 		return this.userService.searchUser(userId);
 	}
 	
 	//http://localhost:9200/user/updateUser/1
 	@RequestMapping(value = "/updateUser/{userId}", method = RequestMethod.PUT, consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-	public User updateUser(@RequestBody User user,@PathVariable Integer userId) {
+	public User updateUser(@RequestBody User user,@PathVariable Long userId) {
 		User newUser = this.userService.searchUser(userId);
 		newUser.setUserName(user.getUserName());
 		newUser.setPassword(user.getPassword());
